Additional changes needed on top of initial sync
