+++ webapp/package.json
@@ -56,6 +56,7 @@
     "jest-canvas-mock": "2.2.0",
     "jest-junit": "11.0.1",
     "mattermost-webapp": "github:mattermost/mattermost-webapp#23f5f93d9f12a7e2b5623e5cee6814366abd9a0f",
+    "node-sass": "4.14.1",
     "sass-loader": "9.0.2",
     "style-loader": "1.2.1",
     "typescript": "3.9.6",
