+++ webapp/webpack.config.js
@@ -61,7 +61,7 @@ module.exports = {
                 },
             },
             {
-                test: /\.(scss)$/,
+                test: /\.(scss|css)$/,
                 use: [
                     'style-loader',
                     {
